JXR-135 Some minor code refactoring
