+++ maven-jxr/src/main/java/org/apache/maven/jxr/DirectoryIndexer.java
@@ -243,12 +243,12 @@ public class DirectoryIndexer
         Iterator<Map<String, ?>> iter = ( (Map) info.get( "allPackages" ) ).values().iterator();
         while ( iter.hasNext() )
         {
-            Map pkgInfo = (Map) iter.next();
+            Map pkgInfo = iter.next();
 
             VelocityContext subContext = new VelocityContext( context );
             subContext.put( "pkgInfo", pkgInfo );
 
-            String outDir = root + '/' + (String) pkgInfo.get( "dir" );
+            String outDir = root + '/' + pkgInfo.get( "dir" );
             doVelocity( "package-summary", outDir, subContext, engine );
             doVelocity( "package-frame", outDir, subContext, engine );
         }
